Correctly frees vstr memory and prevent unaligned memory access

+++ core/src/apps/mimblewimble_coin/finish_decrypting_slate.py
@@ -22,7 +22,7 @@ async def finish_decrypting_slate(context: Context, message: MimbleWimbleCoinFin
 	from storage.cache import delete, get_memory_view, APP_MIMBLEWIMBLE_COIN_ENCRYPTION_AND_DECRYPTION_CONTEXT, APP_MIMBLEWIMBLE_COIN_TRANSACTION_CONTEXT
 	from trezor.wire import NotInitialized, ProcessError, DataError, InvalidSession
 	from trezor.crypto import mimblewimble_coin
-	from uctypes import struct, addressof, INT
+	from uctypes import struct, addressof, UINT8
 	from .coins import getCoinInfo
 	from .common import getExtendedPrivateKey
 	
@@ -50,7 +50,7 @@ async def finish_decrypting_slate(context: Context, message: MimbleWimbleCoinFin
 	encryptionAndDecryptionContextStructure = struct(addressof(encryptionAndDecryptionContext), {
 	
 		# Decrypting state
-		"decryptingState": mimblewimble_coin.ENCRYPTION_AND_DECRYPTION_CONTEXT_DECRYPTING_STATE_OFFSET | INT
+		"decryptingState": mimblewimble_coin.ENCRYPTION_AND_DECRYPTION_CONTEXT_DECRYPTING_STATE_OFFSET | UINT8
 	})
 	
 	# Check if tag is invalid
